Move rotation animation to Surface Runner
Move the rotation animation to the SurfaceAnimationRunner.

The SurfaceControl showing the Screenshot and the actual display are
animated in separate SurfaceAnimator but keep sharing the same state
held in ScreenRotationAnimation.

The SurfaceRotationAnimationHandler delegates the actual transfromations
to the legacy ScreenRotationAnimation and holds the classes needed to
run the animation on the SurfaceAnimationRunner.

A follow up CL will remove references of ScreenRotationAnimation from
the WindowAnimator.

Test: atest DisplayRotationTest, manual test
Bug: 131727512
Change-Id: I8771a0e0eaf41a53411e12d5b4e42a993180d196