Fix resumeWhilePausing not working
Activity was waiting for previous activity to be paused before
restarting, while the activity has FLAG_RESUME_WHILE_PAUSING set.
This behavior was broken after 23866995 if the next top activity does
not have process attached.  Ideally, the next top activity should be
restarted after process was bound, but it was postponed to wait for
all pausing activity completed.

Bring the check (whether need to immediately pause previous activity)
back and move it into AS#startPausing().

Bug: 120003748
Test: atest ActivityLifecycleTests PinnedStackTests ActivityStackTests
Change-Id: I067cb1ea6095816bb573a982c5b50605d1b6d077