Add test mappings to ensure settings backup tests run on appropriate changes.
In core/java/android/provider we want to run the tests which were being run
before they were moved into the SettingsProviderTest suite. To do this I've
followed the existing pattern of running tests marked with the @Presubmit
annotation.

In packages/SettingsProvider we want to run the whole suite on any changes,
which is why there is no filter on that change.

Bug: 139449903
Test: atest and manual verification the moved tests run
Change-Id: If7fd3978a96b979e1cf340803f5a36412de36f65