Use framework_native_aidl for AIDL files in frameworks/native/aidl
The AIDL files under frameworks/native/aidl have been referenced via
their absolute paths. This required any module that has an AIDL file
depending on the AIDL types in the frameworks directory to explicitly
set the include path.

Fixing the problem by abstracting the AIDL files using filegroup where
the path property is set to the base directory for the files. The base
directory is used as include paths when the filegroup is added to srcs.

Bug: 135922046
Test: m
Test: m on sdk_phone_arm7
Change-Id: I8acb39522541f1e06d1a4b69b7a53b328c202eff