Move NotifEntry smart reply/actions to getters
The final piece of our quest to delegate to the Ranking whenever
possible. This one is a bit more complicated since we need to:
(a) Change both types to be nullable (previously they were guaranteed to
be non-null). Most existing code actually functioned under the
assumption they were nullable so this wasn't a huge deal.
(b) Change the smart reply type from CharSequence[] to
List<CharSequence> (to match that of the base Ranking object). This
involves a lot of piping adjustment.

Test: atest
Change-Id: I8b9e33b397606b907103197c54a123a89bff5ad5