Move state changing methods to the Rollback class.
This moves functionality around making rollbacks available, and committing
rollbacks, into the Rollback class. This allows access to the state itself
to be made private within Rollback.

This is part of a series of changes to move rollback-specific functionality
into the Rollback class. Once these are completed, synchronisation can be
moved to be internal to Rollback, and the internal state of the Rollback class
can be cleaned up.

Bug: 136241838
Test: atest RollbackTest
Test: atest StagedRollbackTest
Test: atest CtsRollbackManagerHostTestCases
Test: atest RollbackUnitTest
Change-Id: Ie758cc53b121ce2a84be5ac69641e7c495d76352