Notify display is empty even if it's not the expanded bubble
onSingleTaskDisplayEmpty occurs when the activity is removed from the
activity view (e.g. app crash). When a bubble'd app crashes we want to
collapse the bubble (if expanded) & notify that it's empty so that next
time it's opened it will try to populate the activity again.

We were only doing this if the crashed app was the currently expanded
bubble, but that's not always the case.

This CL adds a method in BubbleData to clean up the display for a given
displayId & calls that in the non-expanded case.

Also adds some null pointer checks.

Test: manual - modify bubbles test app to have a crash when you click a button
             - post some bubbles
             - expand / collapse a bubble
             - crash the bubbles test app
             => see crash dialog
             - expand bubble
             => bubble should be populated
Bug: 138461119
Change-Id: I2fb226766676226980c67dc0536ed37eb43cedf1