Don't mark terminal staged sessions as failed
Currently we mark sessions in an already-terminal state (applied
or failed) as failed when there is a build fingerprint change. This
can cause issues with trying to clean a non-existent directory.
Change this logic to check the status of the session when marking
as failed.

Test: atest StagedInstallTest
Bug: 145840157
Change-Id: I163a3c6558117455038608e15065e076e2dd82d2
Merged-in: I163a3c6558117455038608e15065e076e2dd82d2
(cherry picked from commit 274d0b106f153308815e12c227d90559bceefafc)